refactor(test): updated test w.r.t code
