--- test/fixtures/.eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  rules: {
-    "no-console": [2],
-  },
-  "plugins": [
-    "react"
-  ]
-}
