--- test/fixtures/base-config.yml
@@ -1,5 +1,0 @@
-rules:
-  no-console:
-    - 2
-  no-alert:
-    - 2
