+++ test/fixtures/eslint.yml
@@ -1,0 +1,5 @@
+extends:
+  ./eslintrc
+rules:
+  bar-rule:
+    - 2
