+++ test/fixtures/eslintrc
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "foo-rule": [2]
+  }
+}
