+++ test/fixtures/no-path/package.json
@@ -3,10 +3,5 @@
   "version": "0.0.0",
   "description": "Simulating a project's package.json.",
   "private": true,
-  "main": "./index.js",
-  "eslintConfig": {
-    "rules": {
-      "no-console": [2]
-    }
-  }
+  "main": "./index.js"
 }
